\name{dist.eucsq}
\alias{dist.eucsq}

\title{
Squared euclidean distance metric. Used for ward clustering.
}
\description{
Squared euclidean distance metric. Used for ward clustering.
}
\usage{
dist.eucsq(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
A numeric matrix.
}
}
\details{
Calculates the squared euclidean distance, which is used in the ward
clustering algorithm.
}
\value{
The squared euclidean distance matrix object.
}
\references{
Todo.
}
\author{
Christian Bender (christian.bender@tron-mainz.de)
}
\examples{
\dontrun{
#todo
}
}
\keyword{ internal }
