\name{extract_feat_pam}
\alias{extract_feat_pam}

\title{
Feature extraction for PAMR bootstrapping results.
}
\description{
Feature extraction for PAMR bootstrapping results.
}
\usage{
extract_feat_pam(cvp, SUBDIR=NULL)
}
\arguments{
  \item{cvp}{
	Results of PAMR CV.
}
  \item{SUBDIR}{
	String describing the output folder.
}
}
\details{
An internal function, which is usually not called directly.
}
\value{
Todo.
}
\references{
Todo.
}
\author{
Christian Bender (christian.bender@tron-mainz.de)
}

\seealso{
Todo.
}
\examples{
\dontrun{
#todo
}
}
\keyword{ internal }
