\name{ward}
\alias{ward}

\title{
Clustering function using ward clustering.
}
\description{
Clustering function using ward clustering. Used in \code{drawheat}.
}
\usage{
ward(d)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{d}{
Distance matrix.
}
}
\keyword{ internal }
